pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "attractor"
version = "0.1.0"
description = "DOT-based AI pipeline orchestration system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "httpx>=0.27",
    "openai>=1.30",
    "anthropic>=0.30",
    "google-generativeai>=0.7",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
attractor = "attractor.cli:main"

[tool.setuptools.packages.find]
include = ["attractor*"]
